Add proto/prysm/wrappers for proto-typed HTR helpers
The hash-tree-root helpers that take concrete proto types (transactions,
fork, checkpoint, withdrawals, deposit and withdrawal requests) force
encoding/ssz to import the generated proto packages, which in turn import
encoding/ssz through consensus-types/primitives. Introduce
proto/prysm/wrappers to break potential loops and sidestep issues that can
arise at codegen time when helper methods reference stale types.
