+++ beacon-chain/light-client/lightclient.go
@@ -14,9 +14,9 @@ import (
 	"github.com/OffchainLabs/prysm/v7/consensus-types/interfaces"
 	light_client "github.com/OffchainLabs/prysm/v7/consensus-types/light-client"
 	"github.com/OffchainLabs/prysm/v7/consensus-types/primitives"
-	"github.com/OffchainLabs/prysm/v7/encoding/ssz"
 	enginev1 "github.com/OffchainLabs/prysm/v7/proto/engine/v1"
 	pb "github.com/OffchainLabs/prysm/v7/proto/prysm/v1alpha1"
+	"github.com/OffchainLabs/prysm/v7/proto/prysm/wrappers"
 	"github.com/OffchainLabs/prysm/v7/runtime/version"
 	"github.com/OffchainLabs/prysm/v7/time/slots"
 	"github.com/pkg/errors"
@@ -487,9 +487,9 @@ func ComputeTransactionsRoot(payload interfaces.ExecutionData) ([]byte, error) {
 		}
 		var transactionsRootArray [32]byte
 		if progressiveExecutionPayloadSSZEnabled(payload) {
-			transactionsRootArray, err = ssz.TransactionsRootProgressive(transactions)
+			transactionsRootArray, err = wrappers.TransactionsRootProgressive(transactions)
 		} else {
-			transactionsRootArray, err = ssz.TransactionsRoot(transactions)
+			transactionsRootArray, err = wrappers.TransactionsRoot(transactions)
 		}
 		if err != nil {
 			return nil, errors.Wrap(err, "could not get transactions root")
@@ -510,9 +510,9 @@ func ComputeWithdrawalsRoot(payload interfaces.ExecutionData) ([]byte, error) {
 		}
 		var withdrawalsRootArray [32]byte
 		if progressiveExecutionPayloadSSZEnabled(payload) {
-			withdrawalsRootArray, err = ssz.WithdrawalSliceRootProgressive(withdrawals)
+			withdrawalsRootArray, err = wrappers.WithdrawalSliceRootProgressive(withdrawals)
 		} else {
-			withdrawalsRootArray, err = ssz.WithdrawalSliceRoot(withdrawals, fieldparams.MaxWithdrawalsPerPayload)
+			withdrawalsRootArray, err = wrappers.WithdrawalSliceRoot(withdrawals, fieldparams.MaxWithdrawalsPerPayload)
 		}
 		if err != nil {
 			return nil, errors.Wrap(err, "could not get withdrawals root")
