+++ consensus-types/blocks/BUILD.bazel
@@ -38,6 +38,7 @@ go_library(
         "//proto/engine/v1:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//proto/prysm/v1alpha1/validator-client:go_default_library",
+        "//proto/prysm/wrappers:go_default_library",
         "//runtime/version:go_default_library",
         "@com_github_libp2p_go_libp2p//core/peer:go_default_library",
         "@com_github_libp2p_go_libp2p_pubsub//partialmessages:go_default_library",
