+++ encoding/ssz/BUILD.bazel
@@ -18,8 +18,6 @@ go_library(
         "//crypto/hash:go_default_library",
         "//crypto/hash/htr:go_default_library",
         "//encoding/bytesutil:go_default_library",
-        "//proto/engine/v1:go_default_library",
-        "//proto/prysm/v1alpha1:go_default_library",
         "@com_github_minio_sha256_simd//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
@@ -30,7 +28,6 @@ go_test(
     name = "go_default_test",
     size = "small",
     srcs = [
-        "export_test.go",
         "hashers_test.go",
         "helpers_test.go",
         "htrutils_fuzz_test.go",
@@ -38,16 +35,13 @@ go_test(
         "merkleize_test.go",
         "progressive_merkleize_test.go",
     ],
-    embed = [":go_default_library"],
     deps = [
-        "//config/features:go_default_library",
+        ":go_default_library",
         "//config/fieldparams:go_default_library",
         "//config/params:go_default_library",
         "//crypto/hash:go_default_library",
-        "//encoding/bytesutil:go_default_library",
-        "//proto/engine/v1:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
-        "//runtime/version:go_default_library",
+        "//proto/prysm/wrappers:go_default_library",
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
