--- encoding/ssz/export_test.go
@@ -1,3 +1,0 @@
-package ssz
-
-var WithdrawalRoot = withdrawalRoot
