+++ testing/endtoend/evaluators/BUILD.bazel
@@ -36,11 +36,11 @@ go_library(
         "//container/slice:go_default_library",
         "//crypto/bls:go_default_library",
         "//encoding/bytesutil:go_default_library",
-        "//encoding/ssz:go_default_library",
         "//encoding/ssz/detect:go_default_library",
         "//genesis:go_default_library",
         "//network/httputil:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
+        "//proto/prysm/wrappers:go_default_library",
         "//runtime/interop:go_default_library",
         "//runtime/version:go_default_library",
         "//testing/endtoend/components:go_default_library",
