+++ testing/util/BUILD.bazel
@@ -65,10 +65,10 @@ go_library(
         "//crypto/rand:go_default_library",
         "//crypto/random:go_default_library",
         "//encoding/bytesutil:go_default_library",
-        "//encoding/ssz:go_default_library",
         "//proto/engine/v1:go_default_library",
         "//proto/eth/v1:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
+        "//proto/prysm/wrappers:go_default_library",
         "//runtime/interop:go_default_library",
         "//runtime/version:go_default_library",
         "//testing/assertions:go_default_library",
